Get rid of DBusMap and use the simpler DictEntry
DBus uses Arrays of DictEntry to represent a key-value map.
I had previously thought that I wouldn't model that as Array<DictEntry>, because of the O(n) lookup that comes with that.
I had thought I would make a DBusMap type and model it directly with a HashMap internally.
The problem however, is that DictEntry's may have any basic type as its key, and one of the basic types is the DBus double, that we model with an f64.
However Rust's f64 type cannot be used as a key in HashMap or BTreeMap, because it doesn't implement Hash, Eq or Ord, for good reason.
At the end of the day, just model this thing as Array<DictEntry> using a Vec underneath, and eventually provide a convenience API for the user to convert this stuff to a HashMap or BTreeMap, for types that arent DBusDouble.